Learned the gist of getting text and outputting it via XHTTPREQ - onprogress, onerror, onload, open
Learned the gist of getting text and outputting it via XHTTPREQ - onprogress, onerror, onload, open.

Instead of refreshing after data on the page gets updated, like adding an item to a cart or decreasing the quantity of an item, AJAX allows the user to  do all this without refreshing the page.